requirements.txt
aioredis==1.3.1
amqp==2.6.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
autopep8==1.5.4
beautifulsoup4==4.9.3
billiard==3.5.0.5
bs4==0.0.1
celery==4.2.0
certifi==2020.6.20
cffi==1.14.3
channels==3.0.2
channels-redis==3.2.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
constantly==15.1.0
cryptography==3.2.1
daphne==3.0.0
defusedxml==0.6.0
Django==3.1.2
django-allauth==0.44.0
django-celery-beat==2.1.0
django-celery-results==2.0.0
django-timezone-field==4.0
djangorestframework==3.12.1
future==0.18.2
gprof2dot==2019.11.30
hiredis==1.1.0
humanize==3.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
Jinja2==2.11.2
kombu==4.6.11
lxml==4.6.2
MarkupSafe==1.1.1
msgpack==1.0.0
oauthlib==3.1.0
Pillow==8.0.1
prometheus-client==0.8.0
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
redis==3.5.3
redis-server==6.0.9
requests==2.24.0
requests-oauthlib==1.3.0
serializers==0.2.4
service-identity==18.1.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
toml==0.10.1
tornado==6.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.10
vine==1.3.0
wcwidth==0.2.5
wget==3.2
zope.interface==5.2.0
